requirements.txt
fastapi==0.109.2
uvicorn==0.24.0
python-dotenv==1.0.0
jira==3.5.2
PyGithub==2.1.1
python-dateutil==2.8.2
pydantic==2.6.1
jinja2==3.1.2
aiohttp==3.11.16
cryptography>=3.4.0
pyjwt[crypto]>=2.4.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-text-splitters==0.2.4
ollama==0.3.0 